Additional Networking changes
I swear DatagramChannel is the worst thing that could possibly exist...
Basically I've been forced into doing a big workaround that I didn't
want to have to do.
Because of the way DatagramChannel works, I now have to potentially
overallocate space into a ByteBuffer to make sure that it captures all
the data that it should.
Maybe I should've stuck with the old IO, even though it used more
threads. Because at least that one worked.