Networking Rewrite... Again
Another networking rewrite! This one comes on the back of me deciding
that the old mode of networking incurred way too much overhead. Sending
each packet separately came with lots of overhead and other issues. Now
we're on a Nagle-inspired path (which is why I turned it off)
aggregating data until the time comes to send it. So, instead of
separate threads sending data, that has been merged into the main engine
flow. Packets are now sent at the end of every tick. Reading them still
happens with its own thread, however.

As you can see, lots of leftover debug code is scattered about. I'll
remove it eventually...

The big issue that arose was the serialization. DO IT WHEN PACKETS ARE
INITIALIZED. If you can't remember that, there's a handy dandy comment
above PacketObject to remind you, and explain why.