Added Logging, Config Files, and LaunchWrapper Functionality
Quite the mouthful. This probably should've been 3 separate commits,
but...
Logging is handled using the default Java Logging libraries. Besides the
default ConsoleHandler, two FileHandlers also exist to help log all the
data.
Config Files and LaunchWrapper functionality were quite interdependent,
and required a change in how the Engine was launched. Now, the Engine
has a dedicated launching method (Which existed in a very old version
but was removed. Welp, it's back.) which handles reading from the
Configuration file. More functionality to come in the future.
You may have noticed the new class LaunchConfig. That is the bread and
butter of these new features. It's quite simple, really. But it makes
everything so much easier.
Of course, the LaunchWrapper is still incomplete. It is only functional
at the moment for non-packaged games, and without full support. But
programming is hard.