Additional Utilities and Config Options; Physics and Geometry Overhaul
The keen eyed among all y'all may notice that there are tons of changes
in this commit. I am lazy and forget to commit.

The utilities are associated with turn-based games, and like all the
rest of the included utilities, they are so basic and will 99% require
more work in the future.

The config options are associated with Connection buffer sizes, and
probably shouldn't even be accessible to the end users, so I'm probably
gonna do something about it later.

The physics and geometry overhaul come with me growing increasingly
frustrated with the fact that I was using a Vector class to represent
points which is a cardinal sin in quite a few places. So instead I
separated it out and created new classes and packages. Of course, it is
well documented because I recently worked on it. And of course I need to
work on documentation of quite a bit of the code as well.