Mathematics and Physics Overhaul, and Various Minor Improvements
In conjunction with those introductory math and physics classes, I have
decided to update the engine to better reflect higher-level math and
physics (though not by much, to be honest).
The Collision Detection system makes use of an O(n^2) method to check
the possible collisions between each and every active entity. No longer,
however. If the two hitboxes are not within a certain distance of each
other, the system will not check whether a collision has occured. It is
up to the Hitbox implementation to make sure that the "circleRadius" is
a good value.
I do intend to allow the physics system to make calculations in ever
finer timesteps. A new configuration line has been added, but the rest
of the implementation is not there just yet. Perhaps I may have to start
a separate thread for physics calculations.
Various improvements have been scattered about, such as the
implementation of collision handlers that can be used to process
collisions between different objects in specific manners, rather than
rely on the old ELASTIC/INELASTIC system as before. In accordance to
this new system, the old one has been removed.
Another utility folder has popped up. This one is for mathematics. At
the moment, it is quite bare, and I do not envision much use of it in
the near future. And, of course, it's being overengineered. But nothing
bad could come out of it, right?
Pathfinding's principles have now been separated from the Tile class.
Now, pathfinding is based upon a Pathfinding Node, which Tile
implements. This does allow for things other than just Tiles to make use
of the pathfinding system, however.
Also, you might spot some random commented out lines everywhere. These
are debug lines, which I may have to reactivate in the future.