Replacement of the obsolete Packet system
It's been a while hasn't it...

This commit marks the final replacement of the previously deprecated
packet system. Okay, maybe not replacement, but definitely overhaul.
Instead of being limited in the ways that packets can be sent by their
super class, they can now be placed into either send queue. The old
PacketTCP and PacketUDP are gone, and the protocol wrappers associated
with them are now just held in the protocl wrapper class. Go figure.

To be honest, I haven't the faintest memory of why they were like that
before. It's been too long, so there probably was a half-decent
explanation somewhere. Still, I think this system is far superior.